Color values for PPM were scaled using integers, so everything ended up lovely 16 color

+++ NetpbmViewer/Readers/PpmAsciiPixelReader.cs
@@ -6,7 +6,7 @@ namespace NetpbmViewer.Readers
 {
     public class PpmAsciiPixelReader : INetpbmPixelReader
     {
-        private int _colorMax;
+        private double _colorMax;
 
         public void SetColorMax(Stream stream)
         {
